fix cls caused by universal site layout.
