+++ blocks/breadcrumb/breadcrumb.css
@@ -2,7 +2,6 @@
     display: flex;
     flex-direction: column;
     width: 100%;
-    margin-bottom: 0.5rem;
 }
 
 .breadcrumb-list {
