+++ blocks/news-slider/news-slider.js
@@ -161,7 +161,7 @@ export default function decorate(block) {
       link.className = 'eyebrow-link';
       link.textContent = item.textContent;
 
-      link.addEventListener('click', function (event) {
+      link.addEventListener('click', (event) => {
         event.preventDefault();
         document.querySelectorAll('.eyebrow-link').forEach((el) => {
           el.classList.remove('active-tab');
