+++ templates/universal/universal.css
@@ -5,10 +5,6 @@ main {
   overflow: hidden;
 }
 
-main .main-content-container {
-  margin-top: 0;
-}
-
 .top-ad-section {
   padding-bottom: 20px;
   border-top: 2px solid #000000;
@@ -32,15 +28,6 @@ main .main-content-container {
 .top-ad-section .top-ad {
   z-index: 1;
   position: relative;
-}
-
-main .section.breadcrumb-container {
-  margin: 0;
-  padding: 0;
-}
-
-main .section.breadcrumb-container .breadcrumb-wrapper {
-  max-width: unset;
 }
 
 .ad-title {
@@ -86,10 +73,6 @@ main .section.breadcrumb-container .breadcrumb-wrapper {
   text-align: center;
 }
 
-main .back-to-top {
-  margin: 0 var(--padding-spacer-x);
-}
-
 main .back-top-top-section-header {
   display: flex;
   align-items: center;
@@ -110,6 +93,10 @@ main .back-top-top-section-header {
 
 .fixed-height {
   height: 626px;
+}
+
+main .breadcrumb {
+  margin-top: .25rem;
 }
 
 @keyframes loading {
@@ -141,42 +128,47 @@ main .back-top-top-section-header {
 }
 
 @media (min-width: 992px) {
-  main .content-and-ads-container {
+  /* avoids CLS before the right ad section moves into place */
+  main > .section {
+    margin-right: 330px;
+    padding-right: var(--padding-spacer-x);
+  }
+
+  main.main-grid {
     display: grid;
     gap: var(--padding-spacer-x);
     grid-template-columns: auto 1fr 330px auto;
     grid-template-rows: auto auto;
     grid-template-areas:
-      '. news news .'
-      '. content ads .';
+      'top top top top'
+      '. content right-ads .'
+      '. to-top to-top .';
   }
 
-  main .content-and-ads-container .news-slider-wrapper {
-    grid-area: news;
-    grid-template: '. content ads .' / auto 1fr 330px auto;
+  main.main-grid .section {
+    padding: 0;
+    margin: 0;
   }
 
-  main .content-and-ads-container .content-section {
+  main .content-section {
     grid-area: content;
-    margin: 0;
     padding: 0;
+    margin: 0;
   }
 
-  main .content-and-ads-container .content-section .section {
-    margin: 0;
-    padding: 0;
-  }
-
-  main .content-and-ads-container .right-ad-section {
-    grid-area: ads;
+  main .right-ad-section {
+    grid-area: right-ads;
     min-height: 700px;
     margin: 0;
     padding: 0;
   }
 
-  main .content-and-ads-container .right-ad-section .section {
-    margin: 0;
-    padding: 0;
+  main .top-section {
+    grid-area: top;
+  }
+
+  main .back-to-top {
+    grid-area: to-top;
   }
 }
